requirements.txt
argparse==1.4.0
numpy==1.22.0
pathlib==1.0.1
h5py==3.6.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu110
torchvision==0.8.2+cu110
torchaudio==0.7.2
pyYAML==6.0